removed signup route from main and into signup.py
